Update GitHub Actions workflow versions
Updated versions for various GitHub Actions workflows to the latest available versions and added Maven secrets to environment variables. The actions for checkout, cache, and artifact upload are now updated to version 4. Moreover, setup for OSSRH (OSS Repository Hosting) username and password was added to the environment variables for better safety and streamlined execution.