Update handling of OSSRH password in .github/workflows/build.yml
Modifications were made to improve the referencing of 'OSSRH_PASSWORD_PARAM' in GitHub Actions workflow. By changing the command from double to single quotation marks, we have now improved the handling of the password by ensuring it is interpreted properly. This change enhances the security and correct usage of sensitive OSSRH password data.