Update Git configuration in Github Actions workflow
The build.yml file has been modified to update the way Git is configured within the Github Actions workflow. Now, not only is an ownership check being performed by adding a safe.directory config, but also the user name and email are set to github-actions to streamline the process and avoid potential conflicts.