Update Maven Central release process in workflow
The build workflow has been updated to utilize the extract_staging_repo_id.sh script. This allows the script to pull the ID of the most recent staging repository before releasing to Maven Central. This improves the reliability of the release process by ensuring we are always releasing the latest version.