Refine extraction and usage of staging repository ID
This commit updates the GitHub build workflow by separating the extraction of the staging repository ID and the release to Maven Central into distinct steps. This increases the efficiency and reliability of the release process, by enabling the direct usage of the extracted staging repository ID in the release step.