Remove 'Add GitHub.com to known hosts' steps and reconfigure git
The redundant steps to add GitHub.com to known hosts and manually create an .ssh directory in build.yml workflows have been removed. In their place, git configuration is updated to use a token for GitHub actions. This provides a more secure and direct way to authenticate connections to GitHub repositories.